refactor: Wrap client in IIFE
