+++ lib/_inspect.js
