+++ lib/cli.js
@@ -21,4 +21,4 @@
  */
 'use strict';
 // ~= NativeModule.require('_debugger').start();
-require('./node-inspect').start();
+require('./_inspect').start();
