+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.9.1",
   "description": "Node Inspect",
   "license": "MIT",
-  "main": "lib/node-inspect.js",
+  "main": "lib/_inspect.js",
   "bin": "cli.js",
   "homepage": "https://github.com/buggerjs/node-inspect",
   "repository": {
